+++ test/widget/chat_attachment_test.dart
@@ -102,6 +102,9 @@ void main() async {
   when(
     graphQlProvider.recentChatsTopEvents(3),
   ).thenAnswer((_) => const Stream.empty());
+  when(
+    graphQlProvider.recentChatsTopEvents(3, archived: true),
+  ).thenAnswer((_) => const Stream.empty());
 
   final StreamController<QueryResult> chatEvents = StreamController();
   when(
@@ -330,6 +333,7 @@ void main() async {
       last: null,
       before: null,
       noFavorite: anyNamed('noFavorite'),
+      archived: anyNamed('archived'),
       withOngoingCalls: anyNamed('withOngoingCalls'),
     ),
   ).thenAnswer((_) => Future.value(RecentChats$Query.fromJson(recentChats)));
@@ -571,6 +575,7 @@ final chatData = {
   'members': {'nodes': [], 'totalCount': 0},
   'kind': 'GROUP',
   'isHidden': false,
+  'isArchived': false,
   'muted': null,
   'directLink': null,
   'createdAt': '2021-12-15T15:11:18.316846+00:00',